Makefile changes
Removed SQLOPTION flag; now using OPTION flag for the SQL precompiler, and it in turn passes some options on to the C and RPG compilers (it doesn't let you specify OPTION parameter for the precompiler and also passed through the COMPILEOPT parameter, it's either one or the other.

Also fixed cleanCDeps macro to remove the *.MBR entry that shows up when compiling an SQLC (the precompiler outputs to QTEMP/*.MBR and hands that off to the C compiler).