Add the ability to specify a character encoding for source files (like UTF-8) and a target CCSID for compiled objects (like 37).
Before compiling, explicitly set CCSID attribute on source file and its dependencies to match what we know its encoding to be (the value specified in the makefile).  This insures that it will be properly converted to EBCDIC by the compiler.  (When files are copied to the IFS, the CCSID attribute gets set to some default, like 819.)

Added CCSID support to cvtsrcpf script, CVTSRCF command and CPP.